pyproject.toml
[tool.poetry]
name = "flaskproject2"
version = "0.1.0"
description = ""
authors = ["Валерий Газаров <v.gazarov@fortus.pro>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
SQLAlchemy-Utils = "^0.38.2"
Flask-APScheduler = "^1.12.3"
python-dotenv = "^0.19.2"
psycopg2 = "^2.9.3"
notifiers = "^1.3.0"
loguru = "^0.6.0"
notion = "^0.0.28"
Flask-Cors = "^3.0.10"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.27.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
